FvwmDX: a module for per-screen window-switching
Add FvwmDX, which is a module for trying to treat desk switching as independent
screens.

This is a module which I've pulled out of the archives.  I'm leaving it here in
case this fleshes out into something which is slightly more kind to FVWM.  At
the moment, there's a number of notable issues with this module:

* Only handles M_DESK, doesn't understand PAGES at all;
* Breaks FvwmPager when switching desks;
* Rather than listening for explicit M_DESK events, a SendToModule command has
  to be given.  This isn't desirable, and as this implements its own copy of
  CMD_GotoPage(), doesn't understand the actual command syntax, so it's not
  equivalent.

Unfortunately, I do not know who the original author is, although I had had to
make quite a few tweaks to the code to get it to play nicely with newer FVWM
since the module was written.